Move clipPolygon to @here/harp-geometry.
'clipPolygon()' is a simple general function for clipping polygon
geometries.
