+++ @here/harp-geometry/lib/ClipPolygon.ts
